[-] MO : Updating the updater for merchants who forget to copy their editorial.xml before updating...

+++ install-dev/php/editorial_update.php
@@ -4,7 +4,7 @@ function editorial_update()
 {
 	/*Table creation*/
 
-	if (Db::getInstance()->getValue('SELECT `id_module` FROM `'._DB_PREFIX_.'module` WHERE `name`="editorial"') AND file_exists(dirname(__FILE__).'/../../modules/editorial/editorial.xml'))
+	if (Db::getInstance()->getValue('SELECT `id_module` FROM `'._DB_PREFIX_.'module` WHERE `name`="editorial"'))
 	{
 		Db::getInstance()->Execute('
 		CREATE TABLE `'._DB_PREFIX_.'editorial` (
@@ -23,23 +23,26 @@ function editorial_update()
 		`body_logo_subheading` varchar(255) NOT NULL,
 		PRIMARY KEY (`id_editorial`, `id_lang`))
 		ENGINE=MyISAM DEFAULT CHARSET=utf8');
-
-		$xml = simplexml_load_file(dirname(__FILE__).'/../../modules/editorial/editorial.xml');
-		Db::getInstance()->Execute('
-		INSERT INTO `'._DB_PREFIX_.'editorial`(`id_editorial`, `body_home_logo_link`) VALUES(1, "'.(isset($xml->body->home_logo_link) ? pSQL($xml->body->home_logo_link) : '').'")');
+		
+		if (file_exists(dirname(__FILE__).'/../../modules/editorial/editorial.xml'))
+		{
+			$xml = simplexml_load_file(dirname(__FILE__).'/../../modules/editorial/editorial.xml');
+			Db::getInstance()->Execute('
+			INSERT INTO `'._DB_PREFIX_.'editorial`(`id_editorial`, `body_home_logo_link`) VALUES(1, "'.(isset($xml->body->home_logo_link) ? pSQL($xml->body->home_logo_link) : '').'")');
 
 		
-		$languages = Db::getInstance()->ExecuteS('SELECT * FROM `'._DB_PREFIX_.'lang`');
-		foreach ($languages as $language)
-			Db::getInstance()->Execute('
-			INSERT INTO `'._DB_PREFIX_.'editorial_lang` (`id_editorial`, `id_lang`, `body_title`, `body_subheading`, `body_paragraph`, `body_logo_subheading`)
-			VALUES (1, '.intval($language['id_lang']).', 
-			"'.(isset($xml->body->{'title_'.$language['id_lang']}) ? pSQL($xml->body->{'title_'.$language['id_lang']}) : '').'", 
-			"'.(isset($xml->body->{'subheading_'.$language['id_lang']}) ? pSQL($xml->body->{'subheading_'.$language['id_lang']}) : '').'", 
-			"'.(isset($xml->body->{'paragraph_'.$language['id_lang']}) ? pSQL($xml->body->{'paragraph_'.$language['id_lang']}, true) : '').'", 
-			"'.(isset($xml->body->{'logo_subheading_'.$language['id_lang']}) ? pSQL($xml->body->{'logo_subheading_'.$language['id_lang']}) : '').'")');
+			$languages = Db::getInstance()->ExecuteS('SELECT * FROM `'._DB_PREFIX_.'lang`');
+			foreach ($languages as $language)
+				Db::getInstance()->Execute('
+				INSERT INTO `'._DB_PREFIX_.'editorial_lang` (`id_editorial`, `id_lang`, `body_title`, `body_subheading`, `body_paragraph`, `body_logo_subheading`)
+				VALUES (1, '.intval($language['id_lang']).', 
+				"'.(isset($xml->body->{'title_'.$language['id_lang']}) ? pSQL($xml->body->{'title_'.$language['id_lang']}) : '').'", 
+				"'.(isset($xml->body->{'subheading_'.$language['id_lang']}) ? pSQL($xml->body->{'subheading_'.$language['id_lang']}) : '').'", 
+				"'.(isset($xml->body->{'paragraph_'.$language['id_lang']}) ? pSQL($xml->body->{'paragraph_'.$language['id_lang']}, true) : '').'", 
+				"'.(isset($xml->body->{'logo_subheading_'.$language['id_lang']}) ? pSQL($xml->body->{'logo_subheading_'.$language['id_lang']}) : '').'")');
 
-		unlink(dirname(__FILE__).'/../../modules/editorial/editorial.xml');
+			unlink(dirname(__FILE__).'/../../modules/editorial/editorial.xml');
+		}
 	}
 }
 ?>
